Create a basic CLI and lib division
