+++ package.json
@@ -3,8 +3,9 @@
   "private": true,
   "version": "1.0.0",
   "description": "Optics library for lenses generation",
+  "main": "src/lib.ts",
   "scripts": {
-    "start": "ts-node src/main.ts"
+    "start": "ts-node src/cli.ts"
   },
   "license": "MIT",
   "dependencies": {
