+++ src/cli.ts
@@ -1,0 +1,3 @@
+import {fromFiles} from './lib'
+
+fromFiles(process.argv.slice(2))
